+++ site/src/theme/MDXComponents.js
@@ -16,7 +16,7 @@ import LanguageSwitchLink from '@site/src/components/language/LanguageSwitchLink
 import LanguageSupport from '@site/src/components/language/LanguageSupport';
 import LanguageSwitcher from '@site/src/components/language/LanguageSwitcher';
 import LanguageTabBar from '@site/src/components/language/LanguageTabBar';
-import Dependencies from '@site/src/components/Dependencies';
+import Dependency from '@site/src/components/Dependency';
 import Shnip from '@site/src/components/language/Shnip';
 import CodeBlock from '@theme/CodeBlock';
 const CustomMDXComponents = {
@@ -26,7 +26,7 @@ const CustomMDXComponents = {
   Actors,
   ProjectList,
   Divider,
-  Dependencies,
+  Dependency,
   PillarList,
   SwitchIllustration,
   TwoColumnContainer,
